+++ packages/dn-middleware-rollup/src/getRollupConfig.ts
@@ -30,6 +30,8 @@ export const getRollupConfig = async (opts: IGetRollupConfigOpts, ctx: IDawnCont
     umd,
     esm,
     cjs,
+    system,
+    iife,
     extractCSS = true,
     injectCSS = true,
     cssModules: modules = false,
@@ -371,6 +373,53 @@ export const getRollupConfig = async (opts: IGetRollupConfigOpts, ctx: IDawnCont
               },
             ]
           : []),
+      ];
+    case "system":
+      return [
+        {
+          input,
+          output: {
+            format,
+            sourcemap: (system && system.sourcemap) || false,
+            file: getOutputFile({ entry, type: "system", pkg, bundleOpts }),
+            globals: system && system.globals,
+            name: system && system.name,
+          },
+          plugins: [
+            ...getPlugins({ minCSS: (system && system.minify) || false }),
+            ...extraUmdPlugins,
+            replace({
+              // eslint-disable-next-line @typescript-eslint/naming-convention
+              "process.env.NODE_ENV":
+                system && system.minify ? JSON.stringify("production") : JSON.stringify("development"),
+            }),
+            ...(system && system.minify ? [terser(terserOptions)] : []),
+          ],
+          external: id => testExternal(externalPeerDeps, externalsExclude, id),
+        },
+      ];
+    case "iife":
+      return [
+        {
+          input,
+          output: {
+            format,
+            sourcemap: (iife && iife.sourcemap) || false,
+            file: getOutputFile({ entry, type: "iife", pkg, bundleOpts }),
+            globals: iife && iife.globals,
+          },
+          plugins: [
+            ...getPlugins({ minCSS: (iife && iife.minify) || false }),
+            ...extraUmdPlugins,
+            replace({
+              // eslint-disable-next-line @typescript-eslint/naming-convention
+              "process.env.NODE_ENV":
+                iife && iife.minify ? JSON.stringify("production") : JSON.stringify("development"),
+            }),
+            ...(iife && iife.minify ? [terser(terserOptions)] : []),
+          ],
+          external: id => testExternal(externalPeerDeps, externalsExclude, id),
+        },
       ];
     default:
       throw new Error(`Unsupported type ${type}`);