+++ packages/dn-middleware-rollup/src/index.ts
@@ -75,6 +75,38 @@ const handler: Handler<IOpts> = options => {
         ),
       );
     }
+    if (bundleOpts.system) {
+      tasks.push(
+        run(
+          {
+            cwd,
+            type: "system",
+            entry: bundleOpts.entry,
+            watch,
+            bundleOpts,
+            configFile,
+            analysis,
+          },
+          ctx,
+        ),
+      );
+    }
+    if (bundleOpts.iife) {
+      tasks.push(
+        run(
+          {
+            cwd,
+            type: "iife",
+            entry: bundleOpts.entry,
+            watch,
+            bundleOpts,
+            configFile,
+            analysis,
+          },
+          ctx,
+        ),
+      );
+    }
     await Promise.all(tasks);
 
     await next();