+++ packages/dn-middleware-rollup/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.3.0 (2020-06-28)
+
+### Features
+
+- Add support for IIFE output format
+
+## 1.2.0 (2020-06-28)
+
+### Features
+
+- Add support for SystemJS output format
+
 ## 1.0.0 (2020-06-10)
 
 ### Major